requirements.txt
streamlit>=1.29.0
google-genai>=1.0.0
langdetect>=1.0.9
python-dotenv>=1.0.0
